# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AREA Server API
  version: 0.1.0
  description: |-
    The AREA Server API powers the automation platform that links actions and reactions across
    partner services. Use it to onboard users, issue secure sessions, and orchestrate OAuth-based
    integrations.

    ### Highlights
    - Account lifecycle endpoints for registration, verification, and session management.
    - OAuth helpers to initiate and complete third-party connector authorisations.
    - Metadata routes that expose the current backend build, feature flags, and compatibility notes.

    ### Getting started
    1. Create a user via `POST /v1/users` and complete the verification journey.
    2. Authenticate with either password credentials or OAuth connectors to obtain an `area_session` cookie.
    3. Call authenticated endpoints such as `GET /v1/auth/me` with the session cookie attached as a standard cookie header.

    Refer to the environment selector in the sidebar to toggle between production, staging, and local URLs.
    
    ### Rate limits & reliability
    - Default rate limit: **120 requests per minute** per authenticated session.
    - Burst tolerance: short spikes up to 300 requests per minute are accepted for 30 seconds.
    - All endpoints are reachable over HTTPS only; staging and local URLs are for testing purposes.

    ### Support & SLA
    - Incident response via the on-call rotation is available 24/7 through the status portal.
...
  contact:
...
  license:
...
  x-logo:
...
servers:
...
security:
  - sessionAuth: []
paths:
...
  /v1/areas:
    get:
      summary: List automations owned by the current user
      operationId: listAreas
      tags:
        - areas
      responses:
        '200':
          description: Collection of areas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAreasResponse'
    post:
      summary: Create a new automation for the current user
      operationId: createArea
      tags:
        - areas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAreaRequest'
      responses:
        '201':
          description: Automation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '400':
          description: Invalid payload
        '401':
          description: Authentication required
  /v1/areas/{areaId}/execute:
    post:
      summary: Execute area reactions immediately
      operationId: executeArea
      tags:
        - areas
      parameters:
        - name: areaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Execution accepted
        '401':
          description: Authentication required
        '403':
          description: Area owned by another user
        '404':
          description: Area not found
        '500':
          description: Failed to execute reactions
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: area_session
      description: Session cookie issued after authentication.
  schemas:
...
    CreateAreaRequest:
      type: object
      description: Payload used to create a new automation owned by the authenticated user.
      required: [name, action, reactions]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: Human readable name displayed across clients.
        description:
          type: string
          maxLength: 512
          description: Optional summary to distinguish this automation.
        action:
          $ref: '#/components/schemas/CreateAreaAction'
        reactions:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateAreaReaction'
    CreateAreaAction:
      type: object
      description: Configuration of the action component that triggers the automation.
      required: [componentId]
      properties:
        componentId:
          type: string
          format: uuid
          description: Identifier of the action component selected from the catalog.
        name:
          type: string
          maxLength: 128
          description: Optional nickname to show in clients when displaying the component configuration.
        params:
          type: object
          additionalProperties: true
          description: Free-form JSON parameters persisted in the user component configuration.
    CreateAreaReaction:
      type: object
      description: Configuration of a reaction component executed when the automation fires.
      required: [componentId]
      properties:
        componentId:
          type: string
          format: uuid
          description: Identifier of the reaction component selected from the catalog.
        name:
          type: string
          maxLength: 128
          description: Optional nickname to show in clients when displaying the component configuration.
        params:
          type: object
          additionalProperties: true
          description: Free-form JSON parameters persisted in the user component configuration.
    Area:
      type: object
      description: Automation linking an action to one or more reactions for a user.
      required: [id, name, status, createdAt, updatedAt, reactions]
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the automation.
        name:
          type: string
          description: Display name chosen by the user.
        description:
          type: string
          nullable: true
          description: Optional summary supplied by the user.
        status:
          type: string
          description: Lifecycle status (`enabled`, `disabled`, or `archived`).
        createdAt:
          type: string
          format: date-time
          description: Timestamp (UTC) when the automation was created.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp (UTC) of the last update.
        action:
          $ref: '#/components/schemas/AreaAction'
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/AreaReaction'
    ListAreasResponse:
      type: object
      description: Collection wrapper for automations returned to the client.
      required: [areas]
      properties:
        areas:
          type: array
          items:
            $ref: '#/components/schemas/Area'
    AreaAction:
      type: object
      description: Action binding stored for an AREA automation.
      required: [configId, componentId, component]
      properties:
        configId:
          type: string
          format: uuid
          description: Identifier of the user component configuration backing the action.
        componentId:
          type: string
          format: uuid
          description: Identifier of the catalog component associated with the configuration.
        name:
          type: string
          nullable: true
          description: Optional nickname saved alongside the configuration.
        params:
          type: object
          additionalProperties: true
          description: Persisted configuration parameters supplied when creating the AREA.
        component:
          $ref: '#/components/schemas/ComponentSummary'
    AreaReaction:
      type: object
      description: Reaction binding stored for an AREA automation.
      required: [configId, componentId, component]
      properties:
        configId:
          type: string
          format: uuid
          description: Identifier of the user component configuration backing the reaction.
        componentId:
          type: string
          format: uuid
          description: Identifier of the catalog component associated with the configuration.
        name:
          type: string
          nullable: true
          description: Optional nickname saved alongside the configuration.
        params:
          type: object
          additionalProperties: true
          description: Persisted configuration parameters supplied when creating the AREA.
        component:
          $ref: '#/components/schemas/ComponentSummary'
    ComponentSummary:
      type: object
      description: Minimal catalog metadata required by clients to render a component reference.
      required: [id, kind, name, displayName, provider]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [action, reaction]
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional configuration metadata describing expected parameters.
        provider:
          $ref: '#/components/schemas/ServiceProviderSummary'
...
    ServiceProviderSummary:
      type: object
      description: Lightweight representation of a service provider exposing components.
      required: [id, name, displayName]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
...